CORDA-2399: Samples using public TestCordapp API rather than internal one
Also moved the contents of TestCordappUtils.kt to InternalTestUtils.kt to make it more obvious they're internal.
