+++ finance/workflows/src/test/kotlin/net/corda/finance/internal/CashConfigDataFlowTest.kt
@@ -1,13 +1,12 @@
 package net.corda.finance.internal
 
-import net.corda.core.internal.packageName
 import net.corda.core.utilities.getOrThrow
 import net.corda.finance.EUR
 import net.corda.finance.USD
 import net.corda.testing.node.MockNetwork
 import net.corda.testing.node.MockNetworkParameters
 import net.corda.testing.node.MockNodeParameters
-import net.corda.testing.node.internal.cordappWithPackages
+import net.corda.testing.node.internal.FINANCE_WORKFLOWS_CORDAPP
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.After
 import org.junit.Test
@@ -20,9 +19,9 @@ class CashConfigDataFlowTest {
 
     @Test
     fun `issuable currencies read in from cordapp config`() {
-        val node = mockNet.createNode(MockNodeParameters(additionalCordapps = listOf(
-                cordappWithPackages(javaClass.packageName).copy(config = mapOf("issuableCurrencies" to listOf("EUR", "USD")))
-        )))
+        val node = mockNet.createNode(MockNodeParameters(
+                additionalCordapps = listOf(FINANCE_WORKFLOWS_CORDAPP.withConfig(mapOf("issuableCurrencies" to listOf("EUR", "USD"))))
+        ))
         val config = node.startFlow(CashConfigDataFlow()).getOrThrow()
         assertThat(config.issuableCurrencies).containsExactly(EUR, USD)
     }
